polarweb/requirements.txt
pyserial
numpy
euclid
requests
flask
flask-assets
jsmin
pillow
flask-socketio